v12: Add setting of NCAR_NRDG for resolutions without ridge files

+++ gcm_run.j
@@ -550,9 +550,29 @@ if ( $rst_by_face == YES ) then
     #/bin/cp @GWDRSDIR/gwd_internal_c${AGCM_IM}_face_${n} gwd_internal_face_${n}_rst
   #end
 else
-  /bin/rm gwd_internal_rst
-  /bin/cp @GWDRSDIR/gwd_internal_c${AGCM_IM} gwd_internal_rst
+  # We want to look for the existence of a gwd_internal_rst file
+  # as not all resolutions have them (yet). If it exists, we copy
+  # it to the scratch directory. If it doesn't exist, we need to
+  # add "NCAR_NRDG: 0" to the AGCM.rc file to prevent the model from
+  # trying to use it.
+
+  if (-e @GWDRSDIR/gwd_internal_c${AGCM_IM}) then
+    echo "Found gwd_internal_c${AGCM_IM}. Copying to scratch directory"
+    /bin/rm gwd_internal_rst
+    /bin/cp @GWDRSDIR/gwd_internal_c${AGCM_IM} gwd_internal_rst
+  else
+    echo "WARNING: gwd_internal_rst not found. Setting NCAR_NRDG to 0"
+    # Now, if the user has already set an NCAR_NRDG value, we need to
+    # change it to 0. If they haven't set it, we need to add it to the
+    # AGCM.rc file.
+    if ( `grep -c "NCAR_NRDG:" AGCM.rc` == 0 ) then
+      echo "NCAR_NRDG: 0" >> AGCM.rc
+    else
+      sed -i '/NCAR_NRDG:/c\NCAR_NRDG: 0' AGCM.rc
+    endif
+  endif
 endif
+
 @COUPLED /bin/mkdir INPUT
 @COUPLED cp $EXPDIR/RESTART/* INPUT
 
@@ -1259,7 +1279,7 @@ end
 @MOM5     if(! -e $EXPDIR/MOM_Output) mkdir -p $EXPDIR/MOM_Output
 @MOM5     /bin/mv $SCRDIR/$dset.nc $EXPDIR/MOM_Output/$dset.${edate}.nc
 @MOM5  endif
-@MOM5  end 
+@MOM5  end
 @MOM6  foreach dset ( $dsets )
 @MOM6  set num = `/bin/ls -1 $dset.nc | wc -l`
 @MOM6  if($num != 0) then
